refactor import useShowPost hook shared lib index (#139)
* refactor: add home route, replace profile index to [id].tsx

* refactor condition layout

* refactor index.tsx

* refactor import useShowPost hook shared lib index

---------
